+++ PacletInfo.m
@@ -1,20 +1,19 @@
 (* Paclet Info File *)
 
-(* created 2018/08/30*)
+(* created 2018/09/10*)
 
 Paclet[
     Name -> "MaXrd",
-    Version -> "1.0.1",
+    Version -> "1.0.2",
     MathematicaVersion -> "10.3+",
     Description -> "Databse and utilities for crystallography.",
     Creator -> "Stian Penev Ramsnes",
-    (* URL -> "https://github.com/Stianpr20/MaXrd", *)
-    (* Thumbnail -> "Misc/logo.png", *)
-    Extensions -> {
-		{"Documentation",
-		Resources -> {"Guides/MaXrd"},
-		Language -> "English",
-		MainPage -> "Guides/MaXrd"
-		}
-    }
+    Extensions -> 
+        {
+            {"Documentation", Resources -> 
+                {"Guides/MaXrd"}
+            , Language -> "English", MainPage -> "Guides/MaXrd"}
+        }
 ]
+
+