+++ MaXrd/Changelog.md
@@ -1,4 +1,11 @@
-# Mathematica X-ray diffraction package -- change log
+# Mathematica X-Ray Diffraction Package – change log
+
+
+## Version 1.0.2
+- Fixed a bug where `ReflectionList` and `ReciprocalSpaceSimulation` would not work with crystals that stored wavelength as a `Quantity`.
+- Minor updates and changes in the documentation (thanks to Bianca Eifert for pointing out some of them).
+- Added some missing auto-complete suggestions for `$CrystalData` (for the second argument).
+- Added the option `AngleThreshold` to `BraggAngle` and `ReflectionList` for more efficient filtering by Bragg angle.
 
 
 ## Version 1.0.1